moved Appbar to components Folder
